+++ Common/Nodes/NodalPoint3D.cpp
@@ -12,7 +12,7 @@
 	NodalPoin3Dt: Constructors and Destructor
 ********************************************************************************/
 
-NodalPoint3D::NodalPoint3D(long nodeNum,double xPt,double yPt,double zPt)
+NodalPoint3D::NodalPoint3D(int nodeNum,double xPt,double yPt,double zPt)
 		 : NodalPoint(nodeNum)
 {
     x=xPt;
@@ -28,6 +28,6 @@ NodalPoint3D::NodalPoint3D(long nodeNum,double xPt,double yPt,double zPt)
 void NodalPoint3D::PrintNodalPoint(ostream &os)
 {
     char nline[200];
-	sprintf(nline,"%5ld %15.7e %15.7e %15.7e",num,x,y,z);
+	sprintf(nline,"%5d %15.7e %15.7e %15.7e",num,x,y,z);
     os << nline << endl;
 }
